Crate massa_ledger_exports

General description

TODO

Modules

config 🔒

This file defines a configuration structure containing all settings for the ledger system

controller 🔒

error 🔒

This file defines all error types for the ledger system

key 🔒

ledger_changes 🔒

This file provides structures representing changes to ledger entries

ledger_entry 🔒

This file defines the structure representing an entry in the FinalLedger

mapping_grpc 🔒

Structs

DatastoreUpdateDeserializer

Serializer for datastore field of LedgerEntryUpdate

DatastoreUpdateSerializer

Serializer for datastore field of LedgerEntryUpdate

Key

Disk ledger keys representation

KeyDeserializer

Basic key deserializer

KeySerializer

Basic key serializer

LedgerChanges

represents a list of changes to multiple ledger entries

LedgerConfig

Ledger configuration

LedgerEntry

Structure defining an entry associated to an address in the FinalLedger

LedgerEntryDeserializer

Deserializer for LedgerEntry

LedgerEntrySerializer

Serializer for LedgerEntry

LedgerEntryUpdate

represents an update to one or more fields of a LedgerEntry

LedgerEntryUpdateDeserializer

Deserializer for LedgerEntryUpdate

LedgerEntryUpdateSerializer

Serializer for LedgerEntryUpdate

Enums

KeyType

LedgerError

ledger error

Constants

BALANCE_IDENT

BYTECODE_IDENT

DATASTORE_IDENT

VERSION_IDENT

Traits

LedgerController

Functions

datastore_prefix_from_address

Gives the general prefix of the datastore of an address while respecting a provided key prefix