Crate massa_ledger_exports

General description

TODO

Modules

config 🔒

This file defines a configuration structure containing all settings for the ledger system

controller 🔒

error 🔒

This file defines all error types for the ledger system

key 🔒

ledger_changes 🔒

This file provides structures representing changes to ledger entries

ledger_entry 🔒

This file defines the structure representing an entry in the FinalLedger

mapping_grpc 🔒

types 🔒

Provides various tools to manipulate ledger entries and changes happening on them.

Structs

DatastoreUpdateDeserializer

Serializer for datastore field of LedgerEntryUpdate

DatastoreUpdateSerializer

Serializer for datastore field of LedgerEntryUpdate

Key

Disk ledger keys representation

KeyDeserializer

Basic key deserializer

KeySerializer

Basic key serializer

LedgerChanges

represents a list of changes to multiple ledger entries

LedgerChangesDeserializer

LedgerChanges deserializer

LedgerChangesSerializer

LedgerChanges serializer

LedgerConfig

Ledger configuration

LedgerEntry

Structure defining an entry associated to an address in the FinalLedger

LedgerEntryDeserializer

Deserializer for LedgerEntry

LedgerEntrySerializer

Serializer for LedgerEntry

LedgerEntryUpdate

represents an update to one or more fields of a LedgerEntry

LedgerEntryUpdateDeserializer

Deserializer for LedgerEntryUpdate

LedgerEntryUpdateSerializer

Serializer for LedgerEntryUpdate

SetOrKeepDeserializer

SetOrKeepSerializer

SetUpdateOrDeleteDeserializer

SetUpdateOrDeleteSerializer

Enums

KeyType

LedgerError

ledger error

SetOrDelete

Enum representing a set/delete change on a value T

SetOrKeep

represents a set/keep change

SetUpdateOrDelete

Enumeration representing set/update/delete change on a value T

Constants

BALANCE_IDENT

BYTECODE_IDENT

DATASTORE_IDENT

Traits

Applicable

Trait marking a structure that supports another one (V) being applied to it

LedgerController

Functions

datastore_prefix_from_address