Trait massa_ledger_exports::controller::LedgerController

pub trait LedgerController: Send + Sync + Debug {

    // Required methods
    fn apply_changes(
        &mut self,
        changes: LedgerChanges,
        slot: Slot,
        final_state_data: Option<Vec<u8>>
    );
    fn load_initial_ledger(&mut self) -> Result<(), LedgerError>;
    fn get_balance(&self, addr: &Address) -> Option<Amount>;
    fn get_bytecode(&self, addr: &Address) -> Option<Bytecode>;
    fn entry_exists(&self, addr: &Address) -> bool;
    fn get_data_entry(&self, addr: &Address, key: &[u8]) -> Option<Vec<u8>>;
    fn get_datastore_keys(&self, addr: &Address) -> Option<BTreeSet<Vec<u8>>>;
    fn get_ledger_hash(&self) -> Hash;
    fn get_ledger_part(
        &self,
        last_key: StreamingStep<Key>
    ) -> Result<(Vec<u8>, StreamingStep<Key>), ModelsError>;
    fn set_ledger_part(
        &self,
        data: Vec<u8>
    ) -> Result<StreamingStep<Key>, ModelsError>;
    fn reset(&mut self);
    fn set_initial_slot(&mut self, slot: Slot);
    fn get_slot(&self) -> Result<Slot, ModelsError>;
    fn set_final_state_hash(&mut self, data: Vec<u8>);
    fn get_final_state(&self) -> Result<Vec<u8>, ModelsError>;

}

Required Methods

fn apply_changes( &mut self, changes: LedgerChanges, slot: Slot, final_state_data: Option<Vec<u8>> )

Allows applying LedgerChanges to the final ledger

final_state_data should be non-None only if we are storing a final_state snapshot.

fn load_initial_ledger(&mut self) -> Result<(), LedgerError>

Loads ledger from file

fn get_balance(&self, addr: &Address) -> Option<Amount>

Gets the balance of a ledger entry

Returns

The balance, or None if the ledger entry was not found

fn get_bytecode(&self, addr: &Address) -> Option<Bytecode>

Gets a copy of the bytecode of a ledger entry

Returns

A copy of the found bytecode, or None if the ledger entry was not found

fn entry_exists(&self, addr: &Address) -> bool

Checks if a ledger entry exists

Returns

true if it exists, false otherwise.

fn get_data_entry(&self, addr: &Address, key: &[u8]) -> Option<Vec<u8>>

Gets a copy of the value of a datastore entry for a given address.

Arguments

addr: target address

key: datastore key

Returns

A copy of the datastore value, or None if the ledger entry or datastore entry was not found

fn get_datastore_keys(&self, addr: &Address) -> Option<BTreeSet<Vec<u8>>>

Get every key of the datastore for a given address.

Returns

A BTreeSet of the datastore keys

fn get_ledger_hash(&self) -> Hash

Get the current disk ledger hash

fn get_ledger_part( &self, last_key: StreamingStep<Key> ) -> Result<(Vec<u8>, StreamingStep<Key>), ModelsError>

Get a part of the ledger Used for bootstrap Return: Tuple with data and last key

fn set_ledger_part( &self, data: Vec<u8> ) -> Result<StreamingStep<Key>, ModelsError>

Set a part of the ledger Used for bootstrap Return: Last key inserted

fn reset(&mut self)

Reset the ledger

USED FOR BOOTSTRAP ONLY

fn set_initial_slot(&mut self, slot: Slot)

fn get_slot(&self) -> Result<Slot, ModelsError>

fn set_final_state_hash(&mut self, data: Vec<u8>)

fn get_final_state(&self) -> Result<Vec<u8>, ModelsError>