Trait massa_ledger_exports::LedgerController

pub trait LedgerController: Send + Sync + Debug {
    fn apply_changes(&mut self, changes: LedgerChanges, slot: Slot);
    fn get_parallel_balance(&self, addr: &Address) -> Option<Amount>;
    fn get_bytecode(&self, addr: &Address) -> Option<Vec<u8>>;
    fn entry_exists(&self, addr: &Address) -> bool;
    fn get_data_entry(&self, addr: &Address, key: &Hash) -> Option<Vec<u8>>;
    fn has_data_entry(&self, addr: &Address, key: &Hash) -> bool;
    fn get_entire_datastore(&self, addr: &Address) -> BTreeMap<Hash, Vec<u8>>;
    fn get_full_entry(&self, addr: &Address) -> Option<LedgerEntry>;
    fn get_ledger_part(
        &self, 
        last_key: &Option<Vec<u8>>
    ) -> Result<(Vec<u8>, Option<Vec<u8>>), ModelsError>;
    fn set_ledger_part(
        &self, 
        data: Vec<u8>
    ) -> Result<Option<Vec<u8>>, ModelsError>;
}

Required Methods

fn apply_changes(&mut self, changes: LedgerChanges, slot: Slot)

Allows applying LedgerChanges to the final ledger

fn get_parallel_balance(&self, addr: &Address) -> Option<Amount>

Gets the parallel balance of a ledger entry

Returns

The parallel balance, or None if the ledger entry was not found

fn get_bytecode(&self, addr: &Address) -> Option<Vec<u8>>

Gets a copy of the bytecode of a ledger entry

Returns

A copy of the found bytecode, or None if the ledger entry was not found

fn entry_exists(&self, addr: &Address) -> bool

Checks if a ledger entry exists

Returns

true if it exists, false otherwise.

fn get_data_entry(&self, addr: &Address, key: &Hash) -> Option<Vec<u8>>

Gets a copy of the value of a datastore entry for a given address.

Arguments

addr: target address

key: datastore key

Returns

A copy of the datastore value, or None if the ledger entry or datastore entry was not found

fn has_data_entry(&self, addr: &Address, key: &Hash) -> bool

Checks for the existence of a datastore entry for a given address.

Arguments

addr: target address

key: datastore key

Returns

true if the datastore entry was found, or false if the ledger entry or datastore entry was not found

fn get_entire_datastore(&self, addr: &Address) -> BTreeMap<Hash, Vec<u8>>

Returns

A copy of the datastore sorted by key

fn get_full_entry(&self, addr: &Address) -> Option<LedgerEntry>

TODO: remove when API is updated

fn get_ledger_part(
&self,
last_key: &Option<Vec<u8>>
) -> Result<(Vec<u8>, Option<Vec<u8>>), ModelsError>

Get a part of the ledger Used for bootstrap Return: Tuple with data and last key

fn set_ledger_part(&self, data: Vec<u8>) -> Result<Option<Vec<u8>>, ModelsError>

Set a part of the ledger Used for bootstrap Return: Last key inserted