Trait massa_ledger_exports::LedgerController

pub trait LedgerController: Send + Sync + Debug {
    fn apply_changes(&mut self, changes: LedgerChanges, slot: Slot);
    fn load_initial_ledger(&mut self) -> Result<(), LedgerError>;
    fn get_balance(&self, addr: &Address) -> Option<Amount>;
    fn get_bytecode(&self, addr: &Address) -> Option<Vec<u8>>;
    fn entry_exists(&self, addr: &Address) -> bool;
    fn get_data_entry(&self, addr: &Address, key: &[u8]) -> Option<Vec<u8>>;
    fn get_datastore_keys(&self, addr: &Address) -> Option<BTreeSet<Vec<u8>>>;
    fn get_ledger_hash(&self) -> Hash;
    fn get_ledger_part(
        &self,
        last_key: StreamingStep<Vec<u8>>
    ) -> Result<(Vec<u8>, StreamingStep<Vec<u8>>), ModelsError>;
    fn set_ledger_part(
        &self,
        data: Vec<u8>
    ) -> Result<StreamingStep<Vec<u8>>, ModelsError>;
}

Required Methods

fn apply_changes(&mut self, changes: LedgerChanges, slot: Slot)

Allows applying LedgerChanges to the final ledger

fn load_initial_ledger(&mut self) -> Result<(), LedgerError>

Loads ledger from file

fn get_balance(&self, addr: &Address) -> Option<Amount>

Gets the balance of a ledger entry

Returns

The balance, or None if the ledger entry was not found

fn get_bytecode(&self, addr: &Address) -> Option<Vec<u8>>

Gets a copy of the bytecode of a ledger entry

Returns

A copy of the found bytecode, or None if the ledger entry was not found

fn entry_exists(&self, addr: &Address) -> bool

Checks if a ledger entry exists

Returns

true if it exists, false otherwise.

fn get_data_entry(&self, addr: &Address, key: &[u8]) -> Option<Vec<u8>>

Gets a copy of the value of a datastore entry for a given address.

Arguments

addr: target address

key: datastore key

Returns

A copy of the datastore value, or None if the ledger entry or datastore entry was not found

fn get_datastore_keys(&self, addr: &Address) -> Option<BTreeSet<Vec<u8>>>

Get every key of the datastore for a given address.

Returns

A BTreeSet of the datastore keys

fn get_ledger_hash(&self) -> Hash

Get the current disk ledger hash

fn get_ledger_part(
&self,
last_key: StreamingStep<Vec<u8>>
) -> Result<(Vec<u8>, StreamingStep<Vec<u8>>), ModelsError>

Get a part of the ledger Used for bootstrap Return: Tuple with data and last key

fn set_ledger_part(
&self,
data: Vec<u8>
) -> Result<StreamingStep<Vec<u8>>, ModelsError>

Set a part of the ledger Used for bootstrap Return: Last key inserted